\name{chr7.peaks}
\alias{chr7.peaks}
\docType{data}
\title{
  chr7 peaks
}
\description{
  peaks predicted by PeakSegFPOP on chr7:50410631-53200000 in the
  H3K36me3_TDH_immune data set.
}
\usage{data("chr7.peaks")}
\format{
  A data frame with 41 observations on the following 3 variables.
  \describe{
    \item{\code{sample.id}}{a character vector}
    \item{\code{chromStart}}{a numeric vector}
    \item{\code{chromEnd}}{a numeric vector}
  }
}
\keyword{datasets}
