\name{clusterProblems}
\alias{clusterProblems}
\title{clusterProblems}
\description{Compute a set of non-overlapping segmentation problems.}
\usage{clusterProblems(step1.peaks)}
\arguments{
  \item{step1.peaks}{Maximum number of peaks from overlapping segmentation problems.}
}

\value{data.table of non-overlapping segmentation problems.}

\author{Toby Dylan Hocking}





