\name{ConvertModelList}
\alias{ConvertModelList}
\title{ConvertModelList}
\description{Convert a model list from the non-repetitive format that we get
from the C code to the repetitive format that is more useful for
plotting.}
\usage{ConvertModelList(model.list)}
\arguments{
  \item{model.list}{Value of PeakSegJointHeuristic(...).}
}

\value{List of peaks, segments, loss, modelSelection.}

\author{Toby Dylan Hocking}





