\name{demo.profiles}
\alias{demo.profiles}
\docType{data}
\title{
  Demo profiles
}
\description{
  These profiles resulted in negative means in a buggy version of the
  PeakSegJointFaster code.
}
\usage{data("demo.profiles")}
\format{
  A data frame with 51204 observations on the following 5 variables.
  \describe{
    \item{\code{sample.id}}{a character vector}
    \item{\code{sample.group}}{a character vector}
    \item{\code{chromStart}}{a numeric vector}
    \item{\code{chromEnd}}{a numeric vector}
    \item{\code{count}}{a numeric vector}
  }
}
\keyword{datasets}
