\name{getProblems}
\alias{getProblems}
\title{getProblems}
\description{Tile problems over an entire chromosome.}
\usage{getProblems(chrom, min.chromStart, max.chromEnd, bases.per.problem, 
    overlap.count = 3)}
\arguments{
  \item{chrom}{chrom for problem names.}
  \item{min.chromStart}{First base of problems.}
  \item{max.chromEnd}{Last base of problems.}
  \item{bases.per.problem}{each problem has this many bases.}
  \item{overlap.count}{how many problems overlap at any given genome position?}
}



\author{Toby Dylan Hocking}





