\name{H3K27ac.TDH.MMM4}
\alias{H3K27ac.TDH.MMM4}
\docType{data}
\title{
  Some histone ChIP-seq data
}
\description{
 This is used in the package tests.
}
\usage{data("H3K27ac.TDH.MMM4")}
\format{
  A data frame with 29905 observations on the following 5 variables.
  \describe{
    \item{\code{sample.id}}{a factor}
    \item{\code{chrom}}{a character vector}
    \item{\code{chromStart}}{a numeric vector}
    \item{\code{chromEnd}}{a numeric vector}
    \item{\code{count}}{a numeric vector}
  }
}
\keyword{datasets}
