\name{H3K36me3.AM.immune.chunk21}
\alias{H3K36me3.AM.immune.chunk21}
\docType{data}
\title{
  H3K36me3_AM_immune chunk 21
}
\description{
  Should be able to detect a peak in all 21 samples.
}
\usage{data("H3K36me3.AM.immune.chunk21")}
\format{
  A profile data frame with 198142 observations.
}
\source{
  http://cbio.ensmp.fr/~thocking/chip-seq-chunk-db/ data set
  H3K36me3_AM_immune chunk 21:
  http://cbio.ensmp.fr/~thocking/chip-seq-chunk-db/H3K36me3_AM_immune/21/regions.png
}
\keyword{datasets}
