\name{H3K36me3.TDH.other.chunk1}
\alias{H3K36me3.TDH.other.chunk1}
\docType{data}
\title{
  H3K36me3 TDH other chunk 1
}
\description{
  8 ChIP-seq samples, some with peaks in some regions, some without.
}
\usage{data("H3K36me3.TDH.other.chunk1")}
\format{
  named list of 2 data.frames: "count" contains the noisy data,
  "regions" 
  contains
  the labels.
}
\source{
  http://cbio.ensmp.fr/~thocking/chip-seq-chunk-db/
  data set H3K36me3_TDH_other
  chunk 1.
}
\keyword{datasets}
