\name{H3K4me3.PGP.immune.4608}
\alias{H3K4me3.PGP.immune.4608}
\docType{data}
\title{
H3K4me3_PGP_immune, 4608 bases per problem
}
\description{
  The H3K4me3_PGP_immune data set was split into several PeakSegJoint
  segmentation problems, with 4608 bases per problem.
}
\usage{data("H3K4me3.PGP.immune.4608")}
\format{
  A named list of 443 segmenation problems.
  Each list has features and target for testing the
  InterveralRegression* functions.
  peaks and modelSelection are also included.
}
\source{
  http://cbio.ensmp.fr/~thocking/chip-seq-chunk-db/ data set H3K4me3_PGP_immune.
}
\keyword{datasets}
