\name{H3K4me3.PGP.immune.chunk2}
\alias{H3K4me3.PGP.immune.chunk2}
\docType{data}
\title{
  H3K4me3 PGP immune chunk 2
}
\description{
  In these data the heuristic algorithm does not recover a segmentation
  with a peak for all samples.
}
\usage{data("H3K4me3.PGP.immune.chunk2")}
\format{
  A profile data frame with 36760 observations.
}
\source{
  http://cbio.ensmp.fr/~thocking/chip-seq-chunk-db/H3K4me3_PGP_immune/2/regions.png
}
\keyword{datasets}
