\name{H3K4me3.TDH.other.chunk8}
\alias{H3K4me3.TDH.other.chunk8}
\docType{data}
\title{
  H3K4me3_TDH_other chunk 8 subset
}
\description{
  It should be easy to recover a joint peak in at least 8 of the 10 samples.
}
\usage{data("H3K4me3.TDH.other.chunk8")}
\format{
  A profile data frame (sample.id, chromStart, chromEnd, count)
  with 19866 observations.
}
\source{
  http://cbio.ensmp.fr/~thocking/chip-seq-chunk-db/ data set
  H3K4me3_TDH_other, chunk 8.
  http://cbio.ensmp.fr/~thocking/chip-seq-chunk-db/H3K4me3_TDH_other/8/regions.png 
}
\keyword{datasets}
