\name{peak.at.profile.end}
\alias{peak.at.profile.end}
\docType{data}
\title{
  peak at profile end
}
\description{
  data set for which a peak chromEnd was identical to the data chromEnd
  in an initial buggy version of the solver.
}
\usage{data("peak.at.profile.end")}
\format{
  a list of 27 profile data.frames.
}
\keyword{datasets}
