\name{PeakSegJointError}
\alias{PeakSegJointError}
\title{PeakSegJointError}
\description{Compute number of incorrect regions for every PeakSegJoint model.}
\usage{PeakSegJointError(converted, problem.regions)}
\arguments{
  \item{converted}{Result of ConvertModelList.}
  \item{problem.regions}{data.frame of annotated region labels.}
}

\value{List of error.totals, error.regions, modelSelection, target.}

\author{Toby Dylan Hocking}





