\name{problem.joint.predict.many}
\alias{problem.joint.predict.many}
\title{problem joint predict many}
\description{Compute all joint peak predictions for one separate problem.}
\usage{problem.joint.predict.many(prob.dir)}
\arguments{
  \item{prob.dir}{project/problems/problemID}
}

\value{Nothing.}

\author{Toby Dylan Hocking}





