\name{problem.joint.predict}
\alias{problem.joint.predict}
\title{problem joint predict}
\description{Compute peak predictions for a joint problem.}
\usage{problem.joint.predict(joint.model.RData, jointProblem.dir)}
\arguments{
  \item{joint.model.RData}{path/to/joint.model.RData file (which contains an object called joint.model)}
  \item{jointProblem.dir}{problem to predict peaks.}
}

\value{Nothing.}

\author{Toby Dylan Hocking}





