\name{PeakErrorSamples}
\alias{PeakErrorSamples}
\title{PeakErrorSamples}
\description{Compute PeakError for several samples.}
\usage{PeakErrorSamples(peaks, 
    regions)}
\arguments{
  \item{peaks}{data.frame of \code{peaks} with sample.id.}
  \item{regions}{data.frame of annotated region labels with sample.id.}
}

\value{data.frame of error \code{regions} with sample.id.}

\author{Toby Dylan Hocking}





