\name{problem.joint.predict}
\alias{problem.joint.predict}
\title{problem joint predict}
\description{Compute peak predictions for a joint problem.}
\usage{problem.joint.predict(jointProblem.dir)}
\arguments{
  \item{jointProblem.dir}{project/problems/problemID/jointProblems/jointProbID}
}

\value{data.table of predicted peaks, with 5 columns: chrom, chromStart,
chromEnd, name, mean.}

\author{Toby Dylan Hocking}





